src/simulate/requirements.txt
requests==2.18.4
PyYAML==3.12
requests-threads
requests-futures
